+++ src/components/generateRoles.js
@@ -1,0 +1,20 @@
+const fs = require('fs');
+const path = require('path');
+const xmlbuilder = require('xmlbuilder');
+
+function generateRolesXml(outputDir) {
+    const rolesXml = xmlbuilder.create('roles_definition', { encoding: 'UTF-8' })
+    .ele('role', { id: '' })
+        .ele('name', '').up()
+        .ele('shortname', '').up()
+        .ele('nameincourse', '').up()
+        .ele('description', '').up()
+        .ele('sortorder', '').up()
+        .ele('archetype', '').up()
+    .end({ pretty: true });
+    fs.writeFileSync(path.join(outputDir, 'roles.xml'), rolesXml);
+
+    return rolesXml;
+}
+
+module.exports = generateRolesXml;
